requirements.txt
cached_property
networkx
penman>=1.1.0
pytorch-ignite
regex
sacrebleu==1.4.10
smatch
transformers==2.11.0
wandb
PyYAML>=5.1